refactored out classes
